pyproject.toml
[tool.poetry]
name = "scene_generator"
version = "0.1.0"
description = "Periodically manage scene processing queue."
authors = ["Ben Fasoli <benfasoli@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.8.2"
PyYAML = "^5.4.1"
loguru = "^0.5.3"
requests = "^2.26.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
mypy = "^0.902"
pytest = "^5.2"
isort = "^5.9.1"
types-PyYAML = "^5.4.10"
types-requests = "^2.25.9"

[tool.isort]
profile = "black" 
src_paths = ["stiltctl", "tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
